Remove stale postbox cache entries on app startup, update privacy policy to include caching
Previously, the cache was only trimmed when we cached new postbox data. This meant that if the user doesn't add any new postboxes
for 30 days, we wouldn't clear the cache. By doing it on app startup, we massively reduce the chance of stale cached data, as
long as the user opens the app every so often.

Privacy policy has been updated to tell the user that we do cache postcodes, and how long we intend to cache it for