feat: add ability to register new postboxes by dropping pin on a map
This new ability enables users to register new postboxes without enabling location permissions.
Just drop the pin somewhere on the map and it will fetch all postboxes near that pin for you to add.

This also does set a flag on that postbox that marks it as "unverified" - ie: the app cannot verify
that you have visited that postbox.