feat: add postbox age estimate
Age estimate is based on what we know about monarchs reign's as well as the postbox type.
This is an estimate, and not entirely accurate. EG: we know that the first C3R postbox appeared in 2024 even though E2R died in 2022, but
we don't know if there was a similar delay for any previous monarch.

Also relies on user correctly identifying monarch, as well as my information on lifespans of each postbox type being correct